// src/core/form/validation-error/validation-error-id.enum.ts
export enum ValidationErrorId {
  birthDatePattern = 'birthDatePattern',
  expirationLength = 'expirationLength',
  expirationMonthPattern = 'expirationMonthPattern',
  expirationMinMonth = 'expirationMinMonth',
  expirationMaxMonth = 'expirationMaxMonth',
  expirationYearPattern = 'expirationYearPattern',
  expirationMinYear = 'expirationMinYear',
  expirationMaxYear = 'expirationMaxYear',
  couponPattern = 'couponPattern',
  emailPattern = 'emailPattern',
  cardNumberPattern = 'cardNumberPattern',
  required = 'required',
  minLength = 'minLength',
  maxLength = 'maxLength',
  cpfNumberPattern = 'cpfNumberPattern',
  cvvPattern = 'cvvPattern',
  phonePattern = 'phonePattern',
  buildingNumberPattern = 'buildingNumberPattern',
  zipPattern = 'zipPattern',
  dataPattern = 'dataPattern',
}